refactor: isolate error component
