+++ frontend/ailixir/src/components/molecules/index.ts
@@ -8,6 +8,7 @@ export * from './chat-user-message';
 export * from './action-pair';
 export * from './auth-footer-link';
 export * from './document-detail-actions';
+export * from './failure-card';
 export * from './delete-button';
 export * from './document-list-item';
 export * from './document-page-thumbnail';
